+++ app/src/main/AndroidManifest.xml
@@ -9,14 +9,17 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Stop_and_flight">
-        <activity android:name=".select_signup_method"></activity>
-        <activity android:name=".MainActivity">
+
+        <activity android:name=".IntroActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".AppGuide"/>
+        <activity android:name=".MainActivity" />
+        <activity android:name=".select_signup_method"/>
     </application>
 
 </manifest>
