+++ app/src/main/java/com/example/stop_and_flight/IntroActivity.java
@@ -2,7 +2,9 @@ package com.example.stop_and_flight;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
 
 public class IntroActivity extends AppCompatActivity {
 
@@ -10,5 +12,21 @@ public class IntroActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_intro);
+
+        Handler handler = new Handler();
+        handler.postDelayed(new Runnable(){ @Override
+        public void run() {
+            Intent intent = new Intent(getApplicationContext(), AppGuide.class);
+            startActivity(intent); //introActivity 실행 후 AppGuide로
+             finish();
+        }
+             },2000); //2초 후 인트로 실행
+
     }
+    @Override protected void onPause(){
+        super.onPause();
+        finish();
+    }
+
+
 }
