+++ app/src/main/java/com/example/stop_and_flight/select_signup_method.java
@@ -1,0 +1,26 @@
+package com.example.stop_and_flight;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.firebase.ui.auth.AuthUI;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class select_signup_method extends AppCompatActivity {
+    // Choose authentication providers
+    List<AuthUI.IdpConfig> providers = Arrays.asList(
+            new AuthUI.IdpConfig.EmailBuilder().build(),
+            new AuthUI.IdpConfig.GoogleBuilder().build()
+            );
+    // Create and launch sign-in intent
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_select_signup_method);
+    }
+}
